test: move the placeholder-root test out of the network-marked module
tests/resolver/test_uv_resolver.py is marked network at module level, so
the test was deselected in CI and the resolver change stayed uncovered.
It reads a uv.lock from disk and needs no network.
